dump version 0.2
Since I2C and IO is working, a new version is achieved.

TODO:
The functionality of I2C and IO was ensured with a Linux system, but
further extensions are required in the I2C adapter class. Semaphores are
essential for the read and write method if both the SRTM dispatcher and
other tasks want to use an iface.

It is currently only assumed that the SRTM dispatcher with the I2C
service is the only task that uses an I2C iface.