board:board_io: add IRQ and deffered ISR handling
See:
https://www.freertos.org/Documentation/02-Kernel/02-Kernel-features/11-Deferred-interrupt-handling

The IRQHandle follows the standart ISR processing.
Its task is to evaluate the interrupt from the IO-IFace and determine
which pin has triggered the IRQ.

To keep an ISR short, the response to the IRQ is executed in a separate
task that is 'unblocked' and can therefore be interrupted by other IRQ
events.

The priority of the ISR Callback can be determined with
"configTIMER_TASK_PRIORITY"