add suspend feature
The fsimx8ulp architecture supports suspend and wake-up functions via
the BBNSM (on/off button). As soon as an OFF interrupt is triggered, a
keypress event for the "KEY_SUSPEND" button is simulated in Linux,
shutting down the system.

A new ON/OFF event in PD state triggers wake-up via APP_WakeupACore().
At the same time, a power-up button event, "KEY_WAKEUP", is triggered in
Linux. This enables Linux to recognise that the wake-up was intentional.