add Cortex-M Sleep/DeepSleep support
With the help of this patch, we can ensure that the Cortex-M also enters
sleep mode after an APD-PD event.

It is possible to define a wake-up timeout. The Cortex-M can then
independently, i.e. without external influences, put not only itself but
also the APD into the active state.