picocoremx8ulp: change handling of default DBG-UART
The DBG UART on PCore8ULP was usually UART_B.

However, since the new revision allows UART_B to be used internally for
Bluetooth, it is necessary to switch to UART_D.

This patch checks whether UART_B is being used internally. If not,
UART_B remains the standard DBG Cortex-M interface.